+++ app/Models/LinkClick.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Database\Factories\LinkClickFactory;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -9,7 +10,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class LinkClick extends Model
 {
-    /** @use HasFactory<\Database\Factories\LinkClickFactory> */
+    /** @use HasFactory<LinkClickFactory> */
     use HasFactory;
 
     use HasUuids;