+++ app/Models/Subscriber.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Enums\SubscriberStatus;
+use Database\Factories\SubscriberFactory;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -13,7 +14,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Subscriber extends Model
 {
-    /** @use HasFactory<\Database\Factories\SubscriberFactory> */
+    /** @use HasFactory<SubscriberFactory> */
     use HasFactory;
 
     use HasUuids;